Refactor CreateOrUpdateAdjunct to service
Extracted common code that can be shared by batch logic
